+++ utils/request.go
@@ -74,7 +74,7 @@ func SendCompleteScanRequest(scanData *ScanData, newLog *log.Entry) {
 	newLog.Info("Sending complete scan request")
 	client := &http.Client{}
 
-	req, err := http.NewRequest("DELETE", "http://"+scanData.ApiService+"/api/endgame/"+scanData.Meta.ScanId, nil)
+	req, err := http.NewRequest("DELETE", "http://"+scanData.ApiService+"/api/endgame/"+scanData.Meta.ScanId+"?for_complete=true", nil)
 	if err != nil {
 		panic(fmt.Sprintf("Error occurred while creating complete scan request -> %s", err.Error()))
 	}
@@ -85,4 +85,24 @@ func SendCompleteScanRequest(scanData *ScanData, newLog *log.Entry) {
 		return
 	}
 	newLog.Infof("Complete scan request status received -> %s", resp.Status)
+}
+
+// This function can be used to trigger start scan request
+func SendStartScanRequest(scanData *ScanData, newLog *log.Entry) {
+	tempRequest := map[string]string{"status": "RUNNING", "pid": "15"}
+	tempRequestBody, _ := json.Marshal(tempRequest)
+	temp_ := sendStatusChangeRequestStruct{tempRequestBody}
+
+	postBody, _ := json.Marshal(temp_)
+	requestBody := bytes.NewBuffer(postBody)
+	req, _ := http.NewRequest("PATCH", "http://"+scanData.ApiService+"/api/endgame/"+scanData.Meta.ScanId, requestBody)
+
+	req.Header.Set("Content-Type", "application/json")
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		newLog.Errorf(fmt.Sprintf("Error occurred while sending start scan request -> %s", err.Error()))
+		return
+	}
+	newLog.Infof("Start scan request status received -> %s", resp.Status)
 }